# functions/analytics-svc/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "3.0 - FP0"
  title: Health for Clinical Research
  description: .
basePath: /analytics-svc/api/services/
schemes:
  - http
  - https
produces:
  - application/json
paths:
...
  /cohort:
    x-swagger-router-controller: cohort
    get:
      description: Returns all cohorts in the study
      operationId: getAllCohorts
      produces:
        - application/json
      parameters:
        - name: datasetId
          in: query
          description: Id of dataset
          required: true
          type: string
        - name: offset
          in: query
          description: offset value for results
          required: false
          type: number
        - name: limit
          in: query
          description: limit value for results
          required: false
          type: number
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CohortListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    post:
      description: Sends mriquery to patient analytics service to get list of patient from mriquery, extracts all patient ids, then creates cohort definition and cohort entries in database
      operationId: createCohort
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: cohort
          schema:
            type: object
            required:
              - datasetId
              - mriquery
              - name
              - description
              - owner
            properties:
              datasetId:
                type: string
                description: Id of dataset
              mriquery:
                type: string
                description: Filtercard information in bookmark form. This is a binary string encoded in base64
              name:
                type: string
                description: Cohort Name
              description:
                type: string
                description: Cohort Description
              owner:
                type: string
                description: Cohort Owner
              definitionTypeConceptId:
                type: string
                description: Cohort Definition Type Concept Id
              subjectConceptId: 
                type: number
                description: Cohort Definition Subject Concept Id
              syntax: 
                type: string
                description: Cohort Definition Syntax
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    delete:
      description: Deletes a cohort with corresponding incoming cohortId
      operationId: deleteCohort
      produces:
        - application/json
      parameters:
        - name: datasetId
          in: query
          description: Id of dataset
          required: true
          type: string
        - name: cohortId
          in: query
          type: number
          description: ID of cohort to be deleted from database
          required: true
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
...
  /dataset-filter/filter-scopes:
    x-swagger-router-controller: datasetFilter
    get:
      description: Returns dataset filter
      operationId: getFilterScopes
      produces:
        - application/json
      parameters:
        - name: datasetsWithSchema
          in: query
          type: string
          required: true
      responses:
        "200":
          description: success
          schema:
            $ref: "#/definitions/AnalyticsResponse"
        default: 
          description: Error
          schema:
            $ref: "#/definitions/Error"
...
  /alpdb/{databaseType}/database/{tenant}/schema/{schemaName}/exists:
    x-swagger-router-controller: dbsvc
    get:
      description: Check if schema exists
      operationId: checkIfSchemaExists
      produces:
        - application/json
      parameters:
        - name: databaseType
          in: path
          description: Dialect of database (Hana or Postgres)
          required: true
          type: string
        - name: tenant
          in: path
          description: Tenant database name
          required: true
          type: string
        - name: schemaName
          in: path
          description: Name of schema to check for
          required: true
          type: string
      responses:
        "200":
          description: success
          schema: 
            type: boolean
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
  /alpdb/metadata/schemasnapshot:
    x-swagger-router-controller: dbsvc
    get:
      description: Get schema metadata for snapshot
      operationId: getSnapshotSchemaMetadata
      produces:
        - application/json
      parameters:
        - name: datasetId
          in: query
          description: ID of dataset
          required: true
          type: string
      responses:
        "200":
          description: success
          schema:
            $ref: "#/definitions/SchemaMetadataResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  AnalyticsResponse:
    type: object
    properties:
      data:
        type: object
        description: response object
        properties:
          categories:
            type: array
            items:
              $ref: '#/definitions/Category'
          data:
            type: array
            items:
              type: object
            description: items in this array contains properties that are selected in the UI
          measures:
            type: array
            items:
              $ref: '#/definitions/Measure'
  Measure:
    type: object
    properties:
      group:
        type: integer
        description: grouping
      id:
        type: string
        description: y-axis identifier
      name:   
        type: string
        description: ui friendly y-axis name
      type:   
        type: string
        description: data type of the y-axis data
      value:   
        type: string
        description: property name for the y-axis data     
  Category:
    type: object
    properties:
      axis:
        type: integer 
        description: index in the x-axis
      id:
        type: string
        description: x-axis identifier
      name:   
        type: string
        description: ui friendly x-axis name
      order:   
        type: string
        description: sorting order
        enum:
         - ASC
         - DESC
      type:   
        type: string
        description: data type of the x-axis data
      value:   
        type: string
        description: property name for the x-axis data
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string
...
  CohortResponse:
    type: object
    description: Cohort belonging to the study
    required: [id, patientIds, name, description, creationTimestamp, modificationTimestamp, owner]
    properties:
      id:
        type: string
        description: Unique identifies for a cohort
      patientIds:
        type: array
        description: Array of patient IDs belonging to cohort
        items:
          type: string
      name:
        type: string
        description: Name of cohort
      description:
        type: string
        description: Description of cohort
      creationTimestamp:
        type: string
        description: Indicates when the cohort was created
      modificationTimestamp:
        type: string
        description: Indicates when the cohort was modified
      owner:
        type: string
        description: Indicates owner of cohort
  CohortListResponse:
    type: array
    description: Array of cohorts
    items:
      $ref: "#/definitions/CohortResponse"
...
  SchemaMetadataResponse:
      type: object
      description: Response object when retrieving schema metadata for snapshot